requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
fastapi==0.100.0
graphql-core==3.2.3
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pydantic==2.0.2
pydantic_core==2.1.2
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
strawberry-graphql==0.194.4
typing_extensions==4.7.1
uvicorn==0.22.0
